removing main(updated) and organizing the code.
I added a few functions to simplify the code of the main() function. I added degreesPerSecondConverion(), averageOfsettCalc(), and radiansPerSeconConversion(). I also changed the variable's name to be more specific.I also removed the main(updated) and moved all that to the main file. I have all the printf() commented out and decided to leave them there just for troubleshooting purposes.