reorganizing the code
I changed the opening curly brackets to be in line with the statements. I removed the if statement from the for loop of the AverageOffsetCalc() function. I added snake case for variables and  pascal case for the Functions. I didn't touch any libraries yet so I will do that later.