strain gauge code + library
the initial prototype of the strain gauges.

library: ( port of the Arduino library)
https://os.mbed.com/users/megrootens/code/HX711/docs/tip/classHx711.html

note: the code only measures the voltage difference between the terminals of the Wheatstone bridge. Until it becomes clear what equations are going to be used for the strain I haven't implemented any functions to calculate strain. However, I did add some equations that could be used.