Minimum implementation of CAN
Just a basic interrupt-based CAN bus.  I tried to keep as much of the initialization of the CAN bus in the main.c file. The rest of the CAN bus code is placed in app.cpp.

This basic iteration of the CAN bus is a CAN bus based on a RX interrupt that was set on the ioc file. This RX interrupt is triggered when there is a message pending in the FIFO of the CAN bus which consequently calls the callback function. The function is rather basic and just increments a few numbers and then sends that into the CAN bus . The basic idea is kinda like a call and response. When a certain can ID is received (call) by the stm32 board it responds by sending numbers that are incremented by a counter.

Note: I wasn't sure where I should have placed the callback function so I just placed it into the app.cpp file. Additionally, I wanted to have both CAN_TxHeaderTypeDef TxHeader
CAN_RxHeaderTypeDef RxHeader in the main.c file but the callback function kinda used them so they needed to be in scope so i placed them in the app.cpp file.